feat : Stories기능 구현 , 목록 페이징

+++ src/main/java/com/todo/mypage/domain/MemberInfo.java
@@ -1,13 +1,11 @@
 package com.todo.mypage.domain;
 
-import java.sql.Date;
-
 import lombok.Data;
 
 @Data
 public class MemberInfo {
 
-	private Integer id;
+	private Long id;
 	private String rating;
 	private String privacy;
 	private String email;
